refactor(mod-details): extract ModDetailsModal into shared component
Browse was the only caller; moving it out lets Installed open the
same overlay from a card. Also adds an updateAvailable prop that
flips the primary action button to 'Update' when set, so Installed
can signal when a reinstall is actually an update.
